requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1

pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0

scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3

mlflow==2.17.0

pytest==8.3.3

openai==1.53.0

loguru==0.7.2
